+++ api/src/test/java/software/amazon/ai/ndarray/types/ShapeTest.java
@@ -21,7 +21,7 @@ public class ShapeTest {
     @Test
     public void shapeTest() {
         Shape shape = new Shape(1, 3, 224, 224);
-        Assert.assertEquals(shape.getShape(), new int[] {1, 3, 224, 224});
+        Assert.assertEquals(shape.getShape(), new long[] {1, 3, 224, 224});
         Assert.assertThrows(IllegalArgumentException.class, () -> shape.size(-1));
         Assert.assertEquals(shape.size(0, 1), 3);
         Assert.assertEquals(shape.size(), 150528);
